feat: use new cma plain client
BREAKING CHANGE: We use contentful-management v7 now that provides a new plain client. You will
still be able to use makeRequest from the migration context, but if you happen to use the exported
createManagementClient function you will now be served with a version of the plain client (see docs
of CMA SDK).