Export type declarations to support use as library
I'm working on updates to deluan/contentful-migrate right now, and am
finding that I need to recreate a lot of type information that already
exists in this library. In addition to being a lot of work, duplicating
that information makes my work fragile. It can easily become out of sync
when the types inside contentful-migration change.

If you add this flag, the type declarations (.d.ts files) will become
part of the npm package, letting me depend on them without rewriting
them by hand.

In #213, @chrislambe comments:
> It looks like the interface we want is defined [in the source]. Any
> reason we can't just use that instead of duplicating it?

The answer to his question is that the declarations are not made
available in the npm package. This PR will fix his problem as well as
mine.